+++ ndc-lambda-sdk/test/inference/parallel-degree/invalid.ts
@@ -1,0 +1,30 @@
+/**
+ * @paralleldegree 2
+ */
+export function tagUsedOnProcedure(): string {
+  return "";
+}
+
+/**
+ * @paralleldegree garbage
+ * @readonly
+ */
+export function invalidTagValue(): string {
+  return "";
+}
+
+/**
+ * @paralleldegree 0
+ * @readonly
+ */
+export function zeroParallelDegreeValue(): string {
+  return "";
+}
+
+/**
+ * @paralleldegree -1
+ * @readonly
+ */
+export function negativeDegreeValue(): string {
+  return "";
+}