+++ ndc-lambda-sdk/test/inference/parallel-degree/valid.ts
@@ -1,0 +1,14 @@
+/**
+ * @readonly
+ */
+export function withoutTag(): string {
+  return "";
+}
+
+/**
+ * @readonly
+ * @paralleldegree 666
+ */
+export function withTag(): string {
+  return "";
+}